Update description in config to be clearer
The description section is used in the meta HTML tag in the rendered
output, which in turn is used by, among others, search engines to
provide a short description of the pages.  This commit brings the
description more in line with the general "About" and removes the
history part from the description.